database-service/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
python-jose[cryptography]==3.3.0
sqlmodel==0.0.14